remove old shuffle
