+++ efc/src/shuffle.rs
@@ -1,10 +1,12 @@
+use crate::attestation::{Attestation, AttestationDataSSZ};
 use crate::bls::check_pubkey_key_bls;
-use crate::bls_verifier::G1Json;
+use crate::bls_verifier::{convert_point, G1Json, PairingEntry};
 use crate::utils::{ensure_directory_exists, read_from_json_file};
 use crate::validator::{read_validators, ValidatorPlain, ValidatorSSZ};
 use base64::engine::general_purpose::STANDARD;
 use base64::Engine;
 use circuit_std_rs::gnark::emulated::sw_bls12381::g1::*;
+use circuit_std_rs::gnark::emulated::sw_bls12381::g2::{G2AffP, G2};
 use circuit_std_rs::sha256::m31_utils::big_array_add;
 use circuit_std_rs::utils::{register_hint, simple_select};
 use expander_compiler::circuit::ir::hint_normalized::witness_solver;
@@ -122,10 +124,22 @@ declare_circuit!(ShuffleCircuit {
     position_results: [Variable; SHUFFLE_ROUND * VALIDATOR_CHUNK_SIZE],
     position_bit_results: [Variable; SHUFFLE_ROUND * VALIDATOR_CHUNK_SIZE],
     flip_results: [Variable; SHUFFLE_ROUND * VALIDATOR_CHUNK_SIZE],
-    slot: Variable,
+    //attestationdata
+    slot: [Variable; 8],
+    committee_index: [Variable; 8],
+    beacon_beacon_block_root: [Variable; 32],
+    source_epoch: [Variable; 8],
+    target_epoch: [Variable; 8],
+    source_root: [Variable; 32],
+    target_root: [Variable; 32],
+    //attestationhm = hashtog2(attestationdata.signingroot()), a g2 point
+    attestation_hm: [[[Variable; 48]; 2]; 2], //public hm
+    //attestationsig
+    attestation_sig_bytes: [Variable; 96],
+    attestation_sig_g2: [[[Variable; 48]; 2]; 2], //public sig, unmarsalled from attestation_sig_bytes
     aggregation_bits: [Variable; VALIDATOR_CHUNK_SIZE],
     validator_hashes: [[Variable; POSEIDON_HASH_LENGTH]; VALIDATOR_CHUNK_SIZE],
-    aggregated_pubkey: [[Variable; 48]; 2],
+    aggregated_pubkey: [[Variable; 48]; 2], //public public_key
     attestation_balance: [Variable; 8],
     pubkeys_bls: [[[Variable; 48]; 2]; VALIDATOR_CHUNK_SIZE],
     // validators:      [ValidatorSSZ;VALIDATOR_CHUNK_SIZE],
@@ -142,9 +156,11 @@ declare_circuit!(ShuffleCircuit {
 impl ShuffleCircuit<M31> {
     pub fn from_plains(
         &mut self,
-        shuffle_json: ShuffleJson,
+        shuffle_json: &ShuffleJson,
         plain_validators: &[ValidatorPlain],
         pubkey_bls: &[Vec<String>],
+        attestation: &Attestation,
+        pairing_entry: &PairingEntry,
     ) {
         if shuffle_json.committee_indices.len() != VALIDATOR_CHUNK_SIZE {
             panic!("committee_indices length is not equal to VALIDATOR_CHUNK_SIZE");
@@ -166,7 +182,6 @@ impl ShuffleCircuit<M31> {
             self.position_bit_results[i] = M31::from(shuffle_json.position_bit_results[i]);
             self.flip_results[i] = M31::from(shuffle_json.flip_results[i]);
         }
-        self.slot = M31::from(shuffle_json.slot);
 
         //assign validator_hashes
         for i in 0..VALIDATOR_CHUNK_SIZE {
@@ -245,6 +260,58 @@ impl ShuffleCircuit<M31> {
             for (j, withdrawable_epoch_byte) in withdrawable_epoch.iter().enumerate() {
                 self.withdrawable_epoch[i][j] = M31::from(*withdrawable_epoch_byte as u32);
             }
+
+            //assign slot
+            let slot = attestation.data.slot.to_le_bytes();
+            for (j, slot_byte) in slot.iter().enumerate() {
+                self.slot[j] = M31::from(*slot_byte as u32);
+            }
+            //assign committee_index
+            let committee_index = attestation.data.committee_index.to_le_bytes();
+            for (j, committee_index_byte) in committee_index.iter().enumerate() {
+                self.committee_index[j] = M31::from(*committee_index_byte as u32);
+            }
+            //assign beacon_beacon_block_root
+            let beacon_beacon_block_root = attestation.data.beacon_block_root.clone();
+            let beacon_beacon_block_root = STANDARD.decode(beacon_beacon_block_root).unwrap();
+            for (j, beacon_beacon_block_root_byte) in beacon_beacon_block_root.iter().enumerate() {
+                self.beacon_beacon_block_root[j] = M31::from(*beacon_beacon_block_root_byte as u32);
+            }
+            //assign source_epoch
+            let source_epoch = attestation.data.source.epoch.to_le_bytes();
+            for (j, source_epoch_byte) in source_epoch.iter().enumerate() {
+                self.source_epoch[j] = M31::from(*source_epoch_byte as u32);
+            }
+            //assign target_epoch
+            let target_epoch = attestation.data.target.epoch.to_le_bytes();
+            for (j, target_epoch_byte) in target_epoch.iter().enumerate() {
+                self.target_epoch[j] = M31::from(*target_epoch_byte as u32);
+            }
+            //assign source_root
+            let source_root = attestation.data.source.root.clone();
+            let source_root = STANDARD.decode(source_root).unwrap();
+            for (j, source_root_byte) in source_root.iter().enumerate() {
+                self.source_root[j] = M31::from(*source_root_byte as u32);
+            }
+            //assign target_root
+            let target_root = attestation.data.target.root.clone();
+            let target_root = STANDARD.decode(target_root).unwrap();
+            for (j, target_root_byte) in target_root.iter().enumerate() {
+                self.target_root[j] = M31::from(*target_root_byte as u32);
+            }
+            //assign attestation_hm
+            self.attestation_hm[0] = convert_point(pairing_entry.hm.p.x.clone());
+            self.attestation_hm[1] = convert_point(pairing_entry.hm.p.y.clone());
+
+            //assign attestation_sig_bytes
+            let attestation_sig_bytes = attestation.signature.clone();
+            let attestation_sig_bytes = STANDARD.decode(attestation_sig_bytes).unwrap();
+            for (j, attestation_sig_byte) in attestation_sig_bytes.iter().enumerate() {
+                self.attestation_sig_bytes[j] = M31::from(*attestation_sig_byte as u32);
+            }
+            //assign attestation_sig_g2
+            self.attestation_sig_g2[0] = convert_point(pairing_entry.signature.p.x.clone());
+            self.attestation_sig_g2[1] = convert_point(pairing_entry.signature.p.y.clone());
         }
     }
     pub fn from_pubkey_bls(&mut self, committee_indices: Vec<u32>, pubkey_bls: Vec<Vec<String>>) {
@@ -278,9 +345,7 @@ impl GenericDefine<M31Config> for ShuffleCircuit<Variable> {
             *chunk = simple_select(builder, ignore_flag, zero_var, *chunk);
         }
         //flip the indices based on the hashbit
-        let cur_indices = indices_chunk.clone();
-        let mut copy_cur_indices =
-            builder.new_hint("myhint.copyvarshint", &cur_indices, cur_indices.len());
+        let mut copy_cur_indices = indices_chunk.clone();
         for i in 0..SHUFFLE_ROUND {
             let (cur_indices, diffs) = flip_with_hash_bits(
                 builder,
@@ -300,7 +365,6 @@ impl GenericDefine<M31Config> for ShuffleCircuit<Variable> {
             copy_cur_indices =
                 builder.new_hint("myhint.copyvarshint", &cur_indices, cur_indices.len());
         }
-
         //check the final curIndices, should be equal to the shuffleIndex
         for (i, cur_index) in copy_cur_indices
             .iter_mut()
@@ -381,7 +445,54 @@ impl GenericDefine<M31Config> for ShuffleCircuit<Variable> {
                 builder.assert_is_equal(hashbit, self.validator_hashes[index][i]);
             }
         }
+        // attestation
+        let att_ssz = AttestationDataSSZ {
+            slot: self.slot,
+            committee_index: self.committee_index,
+            beacon_block_root: self.beacon_beacon_block_root,
+            source_epoch: self.source_epoch,
+            target_epoch: self.target_epoch,
+            source_root: self.source_root,
+            target_root: self.target_root,
+        };
+        let mut g2 = G2::new(builder);
+        // domain
+        let domain = [
+            1, 0, 0, 0, 187, 164, 218, 150, 53, 76, 159, 37, 71, 108, 241, 188, 105, 191, 88, 58,
+            127, 158, 10, 240, 73, 48, 91, 98, 222, 103, 102, 64,
+        ];
+        let mut domain_var = vec![];
+        for domain_byte in domain.iter() {
+            domain_var.push(builder.constant(*domain_byte as u32));
+        }
+        let att_hash = att_ssz.att_data_signing_root(builder, &domain_var); //msg
+                                                                            //map to hm
+        let (hm0, hm1) = g2.hash_to_fp(builder, &att_hash);
+        let hm_g2 = g2.map_to_g2(builder, &hm0, &hm1);
+        let expected_hm_g2 = G2AffP::from_vars(
+            self.attestation_hm[0][0].to_vec(),
+            self.attestation_hm[0][1].to_vec(),
+            self.attestation_hm[1][0].to_vec(),
+            self.attestation_hm[1][1].to_vec(),
+        );
+        g2.assert_is_equal(builder, &hm_g2, &expected_hm_g2);
+        // unmarshal attestation sig
+        let sig_g2 = g2.uncompressed(builder, &self.attestation_sig_bytes);
+        let expected_sig_g2 = G2AffP::from_vars(
+            self.attestation_sig_g2[0][0].to_vec(),
+            self.attestation_sig_g2[0][1].to_vec(),
+            self.attestation_sig_g2[1][0].to_vec(),
+            self.attestation_sig_g2[1][1].to_vec(),
+        );
+        g2.assert_is_equal(builder, &sig_g2, &expected_sig_g2);
+        g2.ext2.curve_f.check_mul(builder);
+        g2.ext2.curve_f.table.final_check(builder);
+        g2.ext2.curve_f.table.final_check(builder);
+        g2.ext2.curve_f.table.final_check(builder);
+
         g1.curve_f.check_mul(builder);
+        g1.curve_f.table.final_check(builder);
+        g1.curve_f.table.final_check(builder);
         g1.curve_f.table.final_check(builder);
     }
 }
@@ -489,6 +600,52 @@ pub fn aggregate_attestation_public_key<C: Config, B: RootAPI<C>>(
 ) {
     let one_var = builder.constant(1);
     let mut has_first_flag = builder.constant(0);
+    let mut copy_aggregated_pubkey = pub_key[0].clone();
+    has_first_flag = simple_select(builder, validator_agg_bits[0], one_var, has_first_flag);
+    let mut copy_has_first_flag = builder.new_hint("myhint.copyvarshint", &[has_first_flag], 1)[0];
+    for i in 1..validator_agg_bits.len() {
+        let mut aggregated_pubkey = pub_key[0].clone();
+        let tmp_agg_pubkey = g1.add(builder, &copy_aggregated_pubkey, &pub_key[i]);
+        aggregated_pubkey.x = g1.curve_f.select(
+            builder,
+            validator_agg_bits[i],
+            &tmp_agg_pubkey.x,
+            &copy_aggregated_pubkey.x,
+        );
+        aggregated_pubkey.y = g1.curve_f.select(
+            builder,
+            validator_agg_bits[i],
+            &tmp_agg_pubkey.y,
+            &copy_aggregated_pubkey.y,
+        );
+        let no_first_flag = builder.sub(1, copy_has_first_flag);
+        let is_first = builder.and(validator_agg_bits[i], no_first_flag);
+        aggregated_pubkey.x =
+            g1.curve_f
+                .select(builder, is_first, &pub_key[i].x, &aggregated_pubkey.x);
+        aggregated_pubkey.y =
+            g1.curve_f
+                .select(builder, is_first, &pub_key[i].y, &aggregated_pubkey.y);
+        has_first_flag =
+            simple_select(builder, validator_agg_bits[i], one_var, copy_has_first_flag);
+        copy_aggregated_pubkey = g1.copy_g1(builder, &aggregated_pubkey);
+        copy_has_first_flag = builder.new_hint("myhint.copyvarshint", &[has_first_flag], 1)[0];
+    }
+    g1.curve_f
+        .assert_is_equal(builder, &copy_aggregated_pubkey.x, &agg_pubkey.x);
+    g1.curve_f
+        .assert_is_equal(builder, &copy_aggregated_pubkey.y, &agg_pubkey.y);
+}
+
+pub fn aggregate_attestation_public_key2<C: Config, B: RootAPI<C>>(
+    builder: &mut B,
+    g1: &mut G1,
+    pub_key: &[G1Affine],
+    validator_agg_bits: &[Variable],
+    agg_pubkey: &mut G1Affine,
+) {
+    let one_var = builder.constant(1);
+    let mut has_first_flag = builder.constant(0);
     let mut aggregated_pubkey = pub_key[0].clone();
     has_first_flag = simple_select(builder, validator_agg_bits[0], one_var, has_first_flag);
     for i in 1..validator_agg_bits.len() {
@@ -520,7 +677,6 @@ pub fn aggregate_attestation_public_key<C: Config, B: RootAPI<C>>(
     g1.curve_f
         .assert_is_equal(builder, &aggregated_pubkey.y, &agg_pubkey.y);
 }
-
 pub fn generate_shuffle_witnesses(dir: &str) {
     stacker::grow(32 * 1024 * 1024 * 1024, || {
         println!("preparing solver...");
@@ -557,6 +713,10 @@ pub fn generate_shuffle_witnesses(dir: &str) {
         let shuffle_data: Vec<ShuffleJson> = read_from_json_file(&file_path).unwrap();
         let file_path = format!("{}/pubkeyBLSList.json", dir);
         let public_key_bls_list: Vec<Vec<String>> = read_from_json_file(&file_path).unwrap();
+        let file_path = format!("{}/slotAttestationsFolded.json", dir);
+        let attestations: Vec<Attestation> = read_from_json_file(&file_path).unwrap();
+        let file_path = format!("{}/pairing_assignment.json", dir);
+        let pairing_data: Vec<PairingEntry> = read_from_json_file(&file_path).unwrap();
         let end_time = std::time::Instant::now();
         println!(
             "loaed assignment data, time: {:?}",
@@ -566,19 +726,25 @@ pub fn generate_shuffle_witnesses(dir: &str) {
         let mut handles = vec![];
         let plain_validators = Arc::new(plain_validators);
         let public_key_bls_list = Arc::new(public_key_bls_list);
-        let assignments = Arc::new(Mutex::new(vec![None; shuffle_data.len()]));
+        let attestations = Arc::new(attestations);
+        let assignments = Arc::new(Mutex::new(vec![None; shuffle_data.len() / 2]));
+        let pairing_data = Arc::new(pairing_data);
 
-        for (i, shuffle_item) in shuffle_data.into_iter().enumerate() {
+        for (i, shuffle_item) in shuffle_data.into_iter().enumerate().take(1024) {
             let assignments = Arc::clone(&assignments);
             let target_plain_validators = Arc::clone(&plain_validators);
             let target_public_key_bls_list = Arc::clone(&public_key_bls_list);
+            let target_attestations = Arc::clone(&attestations);
+            let pairing_data = Arc::clone(&pairing_data);
 
             let handle = thread::spawn(move || {
                 let mut assignment = ShuffleCircuit::<M31>::default();
                 assignment.from_plains(
-                    shuffle_item,
+                    &shuffle_item,
                     &target_plain_validators,
                     &target_public_key_bls_list,
+                    &target_attestations[i],
+                    &pairing_data[i],
                 );
 
                 let mut assignments = assignments.lock().unwrap();
@@ -606,7 +772,7 @@ pub fn generate_shuffle_witnesses(dir: &str) {
             .collect::<Vec<_>>();
         let assignment_chunks: Vec<Vec<ShuffleCircuit<M31>>> =
             assignments.chunks(16).map(|x| x.to_vec()).collect();
-        println!("assignment_chunks.len:{}", assignment_chunks.len());
+
         let handles = assignment_chunks
             .into_iter()
             .enumerate()
@@ -636,26 +802,42 @@ pub fn generate_shuffle_witnesses(dir: &str) {
     });
 }
 
-#[test]
-fn run_shuffle() {
-    let dir = "./data";
-    let mut hint_registry = HintRegistry::<M31>::new();
-    register_hint(&mut hint_registry);
-    let start_time = std::time::Instant::now();
-    let plain_validators = read_validators(dir);
-    let file_path = format!("{}/shuffle_assignment.json", dir);
-    let shuffle_data: Vec<ShuffleJson> = read_from_json_file(&file_path).unwrap();
-    let file_path = format!("{}/pubkeyBLSList.json", dir);
-    let public_key_bls_list: Vec<Vec<String>> = read_from_json_file(&file_path).unwrap();
+// #[test]
+// fn test_generate_shuffle2_witnesses() {
+//     generate_shuffle_witnesses("./data");
+// }
 
-    let mut assignment = ShuffleCircuit::<M31>::default();
-    assignment.from_plains(
-        shuffle_data[0].clone(),
-        &plain_validators,
-        &public_key_bls_list,
-    );
+// #[test]
+// fn run_shuffle2() {
+//     let dir = "./data";
+//     let mut hint_registry = HintRegistry::<M31>::new();
+//     register_hint(&mut hint_registry);
+//     let plain_validators = read_validators(dir);
+//     let file_path = format!("{}/shuffle_assignment.json", dir);
+//     let shuffle_data: Vec<ShuffleJson> = read_from_json_file(&file_path).unwrap();
+//     let file_path = format!("{}/pubkeyBLSList.json", dir);
+//     let public_key_bls_list: Vec<Vec<String>> = read_from_json_file(&file_path).unwrap();
+//     let file_path = format!("{}/slotAttestationsFolded.json", dir);
+//     let attestations: Vec<Attestation> = read_from_json_file(&file_path).unwrap();
+//     let file_path = format!("{}/pairing_assignment.json", dir);
+//     let pairing_data: Vec<PairingEntry> = read_from_json_file(&file_path).unwrap();
 
-    stacker::grow(32 * 1024 * 1024 * 1024, || {
-        debug_eval(&ShuffleCircuit::default(), &assignment, hint_registry);
-    });
-}
+//     let mut assignment = ShuffleCircuit::<M31>::default();
+//     assignment.from_plains(
+//         &shuffle_data[0],
+//         &plain_validators,
+//         &public_key_bls_list,
+//         &attestations[0],
+//         &pairing_data[0],
+//     );
+//     let file_name = "shuffle.witness";
+//     stacker::grow(32 * 1024 * 1024 * 1024, || {
+//         let compile_result =
+//             compile_generic(&ShuffleCircuit::default(), CompileOptions::default()).unwrap();
+//         compile_result
+//             .witness_solver
+//             .serialize_into(std::fs::File::create(file_name).unwrap())
+//             .unwrap();
+//         debug_eval(&ShuffleCircuit::default(), &assignment, hint_registry);
+//     });
+// }
